requirements.txt
# Core dependencies for the chat bot
langchain
langchain-google-genai
langchain-community
langchain-huggingface
langchain-chroma
langchain-text-splitters
chromadb
sentence-transformers
python-dotenv==1.0.0
requests==2.31.0
